backend/requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
dj-rest-auth==2.1.3
Django==3.1.4
django-allauth==0.44.0
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-richtextfield==1.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
whitenoise==5.2.0
wrapt==1.12.1
